@@ -3,9 +3,9 @@ package com.loconav.configurator;
 import android.content.Context;
 import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v7.app.AppCompatActivity;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.loconav.configurator.sharedetailsfragmants.DeviceChange;
 import com.loconav.configurator.sharedetailsfragmants.NewInstallation;
